Tests jackson read constraints.
