+++ vertx-core/src/test/java/io/vertx/tests/json/jackson/JacksonDatabindTest.java
@@ -19,8 +19,13 @@ import io.vertx.core.ThreadingModel;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.json.Json;
 import io.vertx.core.json.jackson.DatabindCodec;
+import org.junit.Assume;
 import org.junit.Test;
 
+import java.lang.reflect.ParameterizedType;
+import java.lang.reflect.Type;
+import java.nio.charset.StandardCharsets;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -87,5 +92,42 @@ public class JacksonDatabindTest extends JacksonDatabindTestBase {
     } finally {
       om.setConfig(sc);
     }
+  }
+
+  private static final TypeReference<Integer> INTEGER_TYPE_REF = new TypeReference<Integer>() {};
+  private static final TypeReference<Long> LONG_TYPE_REF = new TypeReference<Long>() {};
+  private static final TypeReference<String> STRING_TYPE_REF = new TypeReference<String>() {};
+  private static final TypeReference<Float> FLOAT_TYPE_REF = new TypeReference<Float>() {};
+  private static final TypeReference<Double> DOUBLE_TYPE_REF = new TypeReference<Double>() {};
+  private static final TypeReference<Map<String, Object>> MAP_TYPE_REF = new TypeReference<Map<String, Object>>() {};
+  private static final TypeReference<List<Object>> LIST_TYPE_REF = new TypeReference<List<Object>>() {};
+  private static final TypeReference<Boolean> BOOLEAN_TYPE_REF = new TypeReference<Boolean>() {};
+
+  @Test
+  public void testDecodeValue() {
+    DatabindCodec codec = new DatabindCodec();
+    assertDecodeValue(codec, Buffer.buffer("42"), 42, INTEGER_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("42"), 42L, LONG_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("\"foobar\""), "foobar", STRING_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("3.4"), 3.4f, FLOAT_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("3.4"), 3.4d, DOUBLE_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("{\"foo\":4}"), Collections.singletonMap("foo", 4), MAP_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("[0,1,2]"), Arrays.asList(0, 1, 2), LIST_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("true"), true, BOOLEAN_TYPE_REF);
+    assertDecodeValue(codec, Buffer.buffer("false"), false, BOOLEAN_TYPE_REF);
+  }
+
+  private <T> void assertDecodeValue(DatabindCodec codec, Buffer buffer, T expected, TypeReference<T> ref) {
+    Type type = ref.getType();
+    Class<?> clazz = type instanceof Class ? (Class<?>) type : (Class<?>) ((ParameterizedType) type).getRawType();
+    assertEquals(expected, codec.fromBuffer(buffer, clazz));
+    assertEquals(expected, codec.fromBuffer(buffer, ref));
+    assertEquals(expected, codec.fromString(buffer.toString(StandardCharsets.UTF_8), clazz));
+    assertEquals(expected, codec.fromString(buffer.toString(StandardCharsets.UTF_8), ref));
+    Buffer nullValue = Buffer.buffer("null");
+    assertNull(codec.fromBuffer(nullValue, clazz));
+    assertNull(codec.fromBuffer(nullValue, ref));
+    assertNull(codec.fromString(nullValue.toString(StandardCharsets.UTF_8), clazz));
+    assertNull(codec.fromString(nullValue.toString(StandardCharsets.UTF_8), ref));
   }
 }
